Refactor calling of offer evaluation
