+++ SingularityService/src/test/java/com/hubspot/singularity/mesos/SingularityStartupTest.java
@@ -52,6 +52,7 @@ public class SingularityStartupTest extends SingularitySchedulerTestBase {
 
     startup.checkSchedulerForInconsistentState();
 
+    scheduler.drainPendingQueue();
     resourceOffers();
 
     Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
